优惠券使用范围问题修正

+++ com.yfshop.shop/com.yfshop.shop.api/src/main/java/com/yfshop/shop/service/coupon/result/YfUserCouponResult.java
@@ -43,7 +43,7 @@ public class YfUserCouponResult implements Serializable {
 
     private String pidPath;
 
-    private Long couponId;
+    private Integer couponId;
 
     private String couponTitle;
